// src/app/core/enums/permissions.enum.ts
export enum ObjectType {
  Digital = '1',
  Magazine = '2',
  NetworkCable = '3',
  NetworkRadio = '4',
  NationalAudio = '4',
  NetworkTV = '5',
  Newspaper = '6',
  Other = '7',
  OutofHome = '8',
  SpotCable = '9',
  SpotRadio = '10',
  SpotTV = '11',
  Owner = '100'
};
  
export enum Service {
  Snapshots = '21',
  Diversity = '22',
  QualifiedFCCDiversity = '23',
  NielsenAudioRatings = '24',
  NationalAudio = '25'
};
  
export enum Role {
  Any = '0',
  AllMedia = '1',
  Digital = '2',
  Magazine = '3',
  NetworkCable = '4',
  NetworkRadio = '5',
  NetworkTV = '6',
  Newspaper = '7',
  OtherMedia = '8',
  Outdoor = '9',
  SpotCable = '10',
  SpotRadio = '11',
  SpotTV = '12',
  Diversity = '13',
  Administrator = '14',
  AccountServices = '15',
  MediaMatch = '16'
};